+++ web/src/pages/ComicsPage.jsx
@@ -6,7 +6,7 @@ import ArtOptions from "../components/ArtOptions.jsx";
 import AddSheet, { ByHand, Searching } from "../components/AddSheet.jsx";
 import BarcodeScan from "../components/BarcodeScan.jsx";
 import EbayLink from "../components/EbayLink.jsx";
-import HelpTip from "../components/HelpTip.jsx";
+import HelpTip, { ShelfHelp } from "../components/HelpTip.jsx";
 import { Icon } from "../components/Icons.jsx";
 import { ShuffleButton } from "../components/Shuffle.jsx";
 import { TagChips, TagEditor, TagFilter } from "../components/Tags.jsx";
@@ -424,6 +424,7 @@ export default function ComicsPage() {
         <span className="rail-spacer" />
         <ViewToggle module="comics" />
         {tiles && inlineDensity && <TileDensity module="comics" />}
+        <ShelfHelp noun="an issue" />
       </div>
       {filtersOpen && (
         <div className="filter-sheet">